Backport TypeScript implementation to JavaScript
