+++ opengrok-indexer/src/main/java/org/opengrok/indexer/analysis/javascript/Consts.java
@@ -24,20 +24,21 @@
 
 package org.opengrok.indexer.analysis.javascript;
 
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
 
 /**
- * Holds static hash set containing the JavaScript keywords.
- * 
- * ECMA-262 5.1 Edition June 2011
+ * Holds JavaScript keywords from ECMA-262 10th Edition, June 2019.
  */
-//TODO update to latest ecmascript ... (for angular support)
 public class Consts {
 
-    public static final Set<String> kwd = new HashSet<>();
+    private static final Set<String> kwd = new HashSet<>();
+
+    public static final Set<String> KEYWORDS = Collections.unmodifiableSet(kwd);
+
     static {
-        //constants
+        // literals
         kwd.add("true");
         kwd.add("false");
         kwd.add("null");
@@ -46,6 +47,7 @@ public class Consts {
         kwd.add("Boolean");
         kwd.add("Date");
         kwd.add("Function");
+        kwd.add("Infinity"); // ECMA-262, 10th edition, June 2019
         kwd.add("Math");
         kwd.add("Number");
         kwd.add("Object");
@@ -56,20 +58,26 @@ public class Consts {
         kwd.add("break");
         kwd.add("case");
         kwd.add("catch");
+        kwd.add("class");
+        kwd.add("const");
         kwd.add("continue");
         kwd.add("debugger");
         kwd.add("default");
         kwd.add("delete");
         kwd.add("do");
         kwd.add("else");
+        kwd.add("export");
+        kwd.add("extends");
         kwd.add("finally");
         kwd.add("for");
         kwd.add("function");
         kwd.add("if");
         kwd.add("in");
         kwd.add("instanceof");
+        kwd.add("import");
         kwd.add("new");
         kwd.add("return");
+        kwd.add("super");
         kwd.add("switch");
         kwd.add("this");
         kwd.add("throw");
@@ -79,14 +87,9 @@ public class Consts {
         kwd.add("void");
         kwd.add("while");
         kwd.add("with");
+        kwd.add("yield");
         //future reserved
-        kwd.add("class");
-        kwd.add("const");
         kwd.add("enum");
-        kwd.add("export");
-        kwd.add("extends");
-        kwd.add("import");
-        kwd.add("super");
         //strict future reserved
         kwd.add("implements");
         kwd.add("interface");
@@ -96,10 +99,8 @@ public class Consts {
         kwd.add("protected");
         kwd.add("public");
         kwd.add("static");
-        kwd.add("yield");
     }
 
-    private Consts() {
+    protected Consts() {
     }
-
 }
