+++ opengrok-indexer/src/main/resources/analysis/typescript/TypeScript.lexh
@@ -19,10 +19,23 @@
 
 /*
  * Copyright (c) 2006, 2017, Oracle and/or its affiliates. All rights reserved.
- * Portions Copyright (c) 2017, Chris Fraire <cfraire@me.com>.
+ * Portions Copyright (c) 2017, 2019, Chris Fraire <cfraire@me.com>.
  */
 
-Identifier = [a-zA-Z_$] [a-zA-Z0-9_$]*
+/*
+ * TypeScript 2.7 added support for numeric separators.
+ */
+DecimalDigits = ({DecimalDigit} |
+    {DecimalDigit} {NumericLiteralSeparator} {DecimalDigit})+
+BinaryDigits = ({BinaryDigit} |
+    {BinaryDigit} {NumericLiteralSeparator} {BinaryDigit})+
+OctalDigits = ({OctalDigit} |
+    {OctalDigit} {NumericLiteralSeparator} {OctalDigit})+
+HexDigits = ({HexDigit} |
+    {HexDigit} {NumericLiteralSeparator} {HexDigit})+
+NumericLiteralSeparator = "_"
 
-Number = (0[xX][0-9a-fA-F]+ | 0[bB][01]+ | [0-9]+\.[0-9]+ |
-    [0-9]+) (([eE][+-]?[0-9]+)?[ufdlUFDL]*)?
+File = [a-zA-Z]{FNameChar}* "." ([Jj][Ss] | [Tt][Ss][Xx]? |
+    [Pp][Rr][Oo][Pp][Ee][Rr][Tt][Ii][Ee][Ss] | [Pp][Rr][Oo][Pp][Ss] |
+    [Xx][Mm][Ll] | [Cc][Oo][Nn][Ff] | [Tt][Xx][Tt] | [Hh][Tt][Mm][Ll]? |
+    [Ii][Nn][Ii] | [Dd][Ii][Ff][Ff] | [Pp][Aa][Tt][Cc][Hh])
