+++ opengrok-indexer/src/main/resources/analysis/typescript/TypeScriptProductions.lexh
@@ -22,282 +22,11 @@
  * Portions Copyright (c) 2017, 2019, Chris Fraire <cfraire@me.com>.
  */
 
-/*
- * U+200C	ZERO WIDTH NON-JOINER
- * U+200D	ZERO WIDTH JOINER
- * Do not support \UnicodeEscapeSequence (yet?)
- */
-Identifier = [\p{XID_Start}_$] [\p{XID_Continue}$\u{200C}\u{200D}]*
-
-/*
- * TypeScript 2.7 added support for numeric separators.
- */
-NumericLiteral = ({DecimalLiteral} | {BinaryIntegerLiteral} |
-    {OctalIntegerLiteral} | {HexIntegerLiteral})
-
-DecimalLiteral = ({DecimalIntegerLiteral} "." {ExponentPart}? |
-    {DecimalIntegerLiteral} "." {DecimalDigits} {ExponentPart}? |
-    "." {DecimalDigits} {ExponentPart}? |
-    {DecimalIntegerLiteral} ({ExponentPart} | {BigIntSuffix})?)
-/*
- * OpenGrok relaxes from the strict ECMA definition of
- *     DecimalIntegerLiteral::
- *         0
- *         NonZeroDigit DecimalDigits?
- * and thus does not bother supporting LegacyOctalIntegerLiteral,
- * NonOctalDecimalIntegerLiteral, or LegacyOctalLikeDecimalIntegerLiteral.
- */
-DecimalIntegerLiteral = {DecimalDigits}
-DecimalDigits = ({DecimalDigit} |
-    {DecimalDigit} {NumericLiteralSeparator} {DecimalDigit})+
-DecimalDigit = [0123456789]
-NonZeroDigit = [123456789]
-ExponentPart = {ExponentIndicator} {SignedInteger}
-ExponentIndicator = [eE]
-SignedInteger = ({DecimalDigits} | "+" {DecimalDigits} | "-" {DecimalDigits})
-NumericLiteralSeparator = "_"    // version 2.7
-BigIntSuffix = "n"    // version 3.2
-
-BinaryIntegerLiteral = "0" [bB] {BinaryDigits}
-BinaryDigits = ({BinaryDigit} |
-    {BinaryDigit} {NumericLiteralSeparator} {BinaryDigit})+
-BinaryDigit = [01]
-
-OctalIntegerLiteral = "0" [oO] {OctalDigits}
-OctalDigits = ({OctalDigit} |
-    {OctalDigit} {NumericLiteralSeparator} {OctalDigit})+
-OctalDigit = [01234567]
-
-HexIntegerLiteral = "0" [xX] {HexDigits}
-HexDigits = ({HexDigit} |
-    {HexDigit} {NumericLiteralSeparator} {HexDigit})+
-HexDigit = [0123456789abcdefABCDEF]
-
-File = [a-zA-Z]{FNameChar}* "." ([Tt][Ss][Xx]? |
-    [Pp][Rr][Oo][Pp][Ee][Rr][Tt][Ii][Ee][Ss] | [Pp][Rr][Oo][Pp][Ss] |
-    [Xx][Mm][Ll] | [Cc][Oo][Nn][Ff] | [Tt][Xx][Tt] | [Hh][Tt][Mm][Ll]? |
-    [Ii][Nn][Ii] | [Dd][Ii][Ff][Ff] | [Pp][Aa][Tt][Cc][Hh])
-
-/*
- * STRING: double-quoted literal.
- * REGEXP: heuristically-detected regular expression bounded by "/" characters.
- * COMMENT: comment text bounded by "/*" and "*/" and not nestable per ECMA
- *     documentation.
- * SCOMMENT: comment text that ends at a line terminator.
- * QSTRING: single-quoted literal.
- * TEMPLATE: template literal bounded by "`" characters and supporting
- *     substitution.
- */
-%state STRING REGEXP COMMENT SCOMMENT QSTRING TEMPLATE
-
-%%
 <YYINITIAL> {
 
     // Handle the following TypeScript 2.7 multi-word built-in type specially.
     "unique" {WhspChar}+ "symbol"    {
         chkLOC();
         offerKeyword(yytext());
-    }
-
-    {Identifier}    {
-        chkLOC();
-        String id = yytext();
-        if (offerSymbol(id, 0, false) && returnOnSymbol()) {
-            return yystate();
-        }
-    }
-
-    {NumericLiteral}    {
-        chkLOC();
-        onDisjointSpanChanged(HtmlConsts.NUMBER_CLASS, yychar);
-        offer(yytext());
-        onDisjointSpanChanged(null, yychar);
-    }
-
-    \"    {
-        chkLOC();
-        yypush(STRING);
-        onDisjointSpanChanged(HtmlConsts.STRING_CLASS, yychar);
-        offer(yytext());
-    }
-
-    \'    {
-        chkLOC();
-        yypush(QSTRING);
-        onDisjointSpanChanged(HtmlConsts.STRING_CLASS, yychar);
-        offer(yytext());
-    }
-
-    \`    {
-        chkLOC();
-        yypush(TEMPLATE);
-        onDisjointSpanChanged(HtmlConsts.STRING_CLASS, yychar);
-        offer(yytext());
-    }
-
-    "/*"    {
-        yypush(COMMENT);
-        onDisjointSpanChanged(HtmlConsts.COMMENT_CLASS, yychar);
-        offer(yytext());
-    }
-
-    "//"    {
-        yypush(SCOMMENT);
-        onDisjointSpanChanged(HtmlConsts.COMMENT_CLASS, yychar);
-        offer(yytext());
-    }
-
-    /*
-     * Literal regexps are in conflict with division "/" and are detected
-     * in javascript based on context and when ambiguous, the division has
-     * a higher precedence. We do a best-effort context matching for
-     * preceding "=" (variable), "(" (function call) or ":" (object).
-     */
-    [:=(] {WhspChar}* "/"    {
-        String capture = yytext();
-        yypush(REGEXP);
-        offer(capture.substring(0, capture.length() - 1));
-        onDisjointSpanChanged(HtmlConsts.STRING_CLASS, yychar + yylength() - 1);
-        offer("/");
-    }
-
-    [\{\}]    {
-        chkLOC();
-        String capture = yytext();
-        if (notInTemplateOrSubstitutionDoesNotEnd(capture)) {
-            offer(capture);
-        }
-    }
-}
-
-<STRING> {
-    \\[\"\\]    {
-        chkLOC();
-        offer(yytext());
-    }
-
-    \"    {
-        chkLOC();
-        offer(yytext());
-        onDisjointSpanChanged(null, yychar);
-        yypop();
-    }
-}
-
-<REGEXP> {
-    \\[/]    {
-        chkLOC();
-        onNonSymbolMatched(yytext(), yychar);
-    }
-    \/[gimsuy]*    {
-        chkLOC();
-        offer("/");
-        onDisjointSpanChanged(null, yychar);
-        offer(yytext().substring(1));
-        yypop();
-    }
-}
-
-<QSTRING> {
-    \\[\'\\]    {
-        chkLOC();
-        offer(yytext());
-    }
-
-    \'    {
-        chkLOC();
-        onDisjointSpanChanged(null, yychar);
-        offer(yytext());
-        yypop();
-    }
-}
-
-<TEMPLATE> {
-    \\[\`\$\\]    {
-        chkLOC();
-        offer(yytext());
-    }
-
-    \`    {
-        chkLOC();
-        onDisjointSpanChanged(null, yychar);
-        offer(yytext());
-        yypop();
-    }
-
-    "${"    {
-        chkLOC();
-        offer(yytext());
-        onDisjointSpanChanged(null, yychar);
-        pushData();
-        substitutionOp();
-        yypush(YYINITIAL);
-    }
-}
-
-<COMMENT> {
-    "*/"    {
-        onDisjointSpanChanged(null, yychar);
-        offer(yytext());
-        yypop();
-    }
-}
-
-<SCOMMENT> {
-    {WhspChar}*{EOL}    {
-        onDisjointSpanChanged(null, yychar);
-        yypop();
-        onEndOfLineMatched(yytext(), yychar);
-    }
-}
-
-<YYINITIAL, STRING, REGEXP, COMMENT, SCOMMENT, QSTRING, TEMPLATE> {
-    {WhspChar}*{EOL}    {
-        onEndOfLineMatched(yytext(), yychar);
-    }
-    [[\s]--[\n]]    {
-        onNonSymbolMatched(yytext(), yychar);
-    }
-    [^\n]    {
-        chkLOC();
-        onNonSymbolMatched(yytext(), yychar);
-    }
-}
-
-<STRING, COMMENT, SCOMMENT, QSTRING, TEMPLATE> {
-    {FPath}    {
-        chkLOC();
-        onPathlikeMatched(yytext(), '/', false, yychar);
-    }
-
-    {File}    {
-        chkLOC();
-        String path = yytext();
-        onFilelikeMatched(path, yychar);
-    }
-
-    {FNameChar}+ "@" {FNameChar}+ "." {FNameChar}+    {
-        chkLOC();
-        onEmailAddressMatched(yytext(), yychar);
-    }
-}
-
-<STRING, SCOMMENT, TEMPLATE> {
-    {BrowseableURI}    {
-        chkLOC();
-        onUriMatched(yytext(), yychar);
-    }
-}
-
-<COMMENT> {
-    {BrowseableURI}    {
-        onUriMatched(yytext(), yychar, StringUtils.END_C_COMMENT);
-    }
-}
-
-<QSTRING> {
-    {BrowseableURI}    {
-        chkLOC();
-        onUriMatched(yytext(), yychar, StringUtils.APOS_NO_BSESC);
     }
 }
