+++ opengrok-indexer/src/test/resources/analysis/javascript/sample_xref.html
@@ -114,7 +114,7 @@ function get_sym_list(){return [["Class","xc",[["Array",247],["Heatmap",29],["da
 <a class="l" name="108" href="#108">108</a>
 <a class="l" name="109" href="#109">109</a>		<b>this</b>.<a href="/source/s?defs=data_groups" class="intelliWindow-symbol" data-definition-place="undefined-in-file">data_groups</a> = <b>this</b>.<a href="/source/s?defs=makeDrawAreaComponent" class="intelliWindow-symbol" data-definition-place="undefined-in-file">makeDrawAreaComponent</a>(
 <a class="hl" name="110" href="#110">110</a>			<span class="s">&apos;data-groups&apos;</span>,
-<a class="l" name="111" href="#111">111</a>			`<a href="/source/s?defs=translate" class="intelliWindow-symbol" data-definition-place="undefined-in-file">translate</a>(<span class="n">0</span>, <span class="n">20</span>)`
+<a class="l" name="111" href="#111">111</a>			<span class="s">`translate(0, 20)`</span>
 <a class="l" name="112" href="#112">112</a>		);
 <a class="l" name="113" href="#113">113</a>	}
 <a class="l" name="114" href="#114">114</a>
