+++ opengrok-indexer/src/test/resources/analysis/javascript/sampletags
@@ -3,8 +3,9 @@
 !_TAG_PROGRAM_AUTHOR	Universal Ctags Team	//
 !_TAG_PROGRAM_NAME	Universal Ctags	/Derived from Exuberant Ctags/
 !_TAG_PROGRAM_URL	https://ctags.io/	/official site/
-!_TAG_PROGRAM_VERSION	0.0.0	/b13cb551/
+!_TAG_PROGRAM_VERSION	0.0.0	/f43463a/
 !_TAG_OUTPUT_MODE	u-ctags	/u-ctags or e-ctags/
+!_TAG_OUTPUT_FILESEP	slash	/slash or backslash/
 Heatmap	grok.js	/^export default class Heatmap extends BaseChart {$/;"	class	line:29	inherits:BaseChart
 constructor	grok.js	/^	constructor({$/;"	method	line:30	class:Heatmap	signature:({ start = '', domain = '', subdomain = '', data = {}, discrete_domains = 0 + 3.0 - 3.0, count_label = '', legend_colors = [] })
 validate_colors	grok.js	/^	validate_colors(colors) {$/;"	method	line:65	class:Heatmap	signature:(colors)
@@ -15,10 +16,10 @@ setup_components	grok.js	/^	setup_components() {$/;"	method	line:105	class:Heatm
 setup_values	grok.js	/^	setup_values() {$/;"	method	line:115	class:Heatmap	signature:()
 render_all_weeks_and_store_x_values	grok.js	/^	render_all_weeks_and_store_x_values(no_of_weeks) {$/;"	method	line:129	class:Heatmap	signature:(no_of_weeks)
 get_week_squares_group	grok.js	/^	get_week_squares_group(current_date, index) {$/;"	method	line:157	class:Heatmap	signature:(current_date, index)
-data-date	grok.js	/^				'data-date': getDdMmYyyy(current_date),$/;"	property	line:191	class:get_week_squares_group.dataAttr
-data-value	grok.js	/^				'data-value': data_value,$/;"	property	line:192	class:get_week_squares_group.dataAttr
-data-day	grok.js	/^				'data-day': current_date.getDay()$/;"	property	line:193	class:get_week_squares_group.dataAttr
-dataAttr	grok.js	/^			let dataAttr = {$/;"	class	line:190	class:get_week_squares_group
+data-date	grok.js	/^				'data-date': getDdMmYyyy(current_date),$/;"	property	line:191	class:Heatmap.get_week_squares_group.dataAttr
+data-value	grok.js	/^				'data-value': data_value,$/;"	property	line:192	class:Heatmap.get_week_squares_group.dataAttr
+data-day	grok.js	/^				'data-day': current_date.getDay()$/;"	property	line:193	class:Heatmap.get_week_squares_group.dataAttr
+dataAttr	grok.js	/^			let dataAttr = {$/;"	class	line:190	class:Heatmap.get_week_squares_group
 render_month_labels	grok.js	/^	render_month_labels() {$/;"	method	line:219	class:Heatmap	signature:()
 make_graph_components	grok.js	/^	make_graph_components() {$/;"	method	line:246	class:Heatmap	signature:()
 Array	grok.js	/^		Array.prototype.slice.call($/;"	class	line:247
